+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_models.py
@@ -205,9 +205,9 @@ class RecognizePiiEntitiesResult(DictMixin):
             )[:1024]
 
 
-class AnalyzeHealthcareEntitiesResultItem(DictMixin):
+class AnalyzeHealthcareEntitiesResult(DictMixin):
     """
-    AnalyzeHealthcareEntitiesResultItem contains the Healthcare entities from a
+    AnalyzeHealthcareEntitiesResult contains the Healthcare entities from a
     particular document.
 
     :ivar str id: Unique, non-empty document identifier that matches the
@@ -230,7 +230,7 @@ class AnalyzeHealthcareEntitiesResultItem(DictMixin):
     :vartype statistics:
         ~azure.ai.textanalytics.TextDocumentStatistics
     :ivar bool is_error: Boolean check for error item when iterating over list of
-        results. Always False for an instance of a AnalyzeHealthcareEntitiesResultItem.
+        results. Always False for an instance of a AnalyzeHealthcareEntitiesResult.
     """
 
     def __init__(self, **kwargs):
@@ -256,7 +256,7 @@ class AnalyzeHealthcareEntitiesResultItem(DictMixin):
         )
 
     def __repr__(self):
-        return "AnalyzeHealthcareEntitiesResultItem(id={}, entities={}, entity_relations={}, warnings={}, "\
+        return "AnalyzeHealthcareEntitiesResult(id={}, entities={}, entity_relations={}, warnings={}, "\
         "statistics={}, is_error={})".format(
             self.id,
             repr(self.entities),